* Option to reduce transformation when parsing and print formating * Support for binary and other prefixes * Information units (bit, byte, etc.) and binary prefixes * Neper and Bel units

+++ libqalculate/Function.cc
@@ -327,7 +327,7 @@ int MathFunction::args(const string &argstr, MathStructure &vargs, const ParseOp
 		unended_function->set(vargs);
 		unended_function->setType(STRUCT_FUNCTION);
 		unended_function->setFunction(this);
-		while(unended_function->size() < itmp) {
+		while((int) unended_function->size() < itmp) {
 			unended_function->addChild(m_undefined);
 		}
 	}
